backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
PyJWT==2.10.1
bcrypt==4.2.1
APScheduler==3.10.4
alembic==1.14.1
